% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test0620.R
\name{ESizeCoHedg}
\alias{ESizeCoHedg}
\title{Calculating the Effect Size of Cohen's d & Hedges' g}
\usage{
ESizeCoHedg()
}
\value{
Effect Size of Cohen's d & Hedges' g
}
\description{
\code{ESizeCoHedg} calculate the Effect Size & draw the dnorm curve
}
\examples{
# ESizeCoHedg()
}
